feat(map): swipe and center for testing
