+++ packages/geo/src/lib/import-export/import-export/import-export.component.ts
@@ -46,7 +46,7 @@ import type { WorkspaceStore } from '@igo2/common';
 import { WfsWorkspace } from '../../workspace/shared/wfs-workspace';
 import { FeatureWorkspace } from '../../workspace/shared/feature-workspace';
 import { MatSlideToggleChange } from '@angular/material/slide-toggle';
-import { InputProjections, ProjectionsLimitationsOptions } from './import-export.interface';
+import { InputProjections, ProjectionsLimitationsOptions } from '../../map/';
 import { DownloadService } from '../../download/shared/download.service';
 
 @Component({
