+++ packages/geo/src/lib/import-export/import-export/index.ts
@@ -1,2 +1,1 @@
 export * from './import-export.component';
-export * from './import-export.interface';
