+++ packages/geo/src/lib/map/map-center/index.ts
@@ -1,2 +1,1 @@
 export * from './map-center.component';
-export * from './shared/map-center.utils';
